simplify lockout functionality

+++ backend/platform/shared/constants.rs
@@ -1,6 +1,6 @@
 pub mod auth {
     /// Maximum allowed failed login attempts within the defined window before account is temporarily locked.
-    pub const FAILED_LOGIN_MAX_ATTEMPTS: i64 = 5;
+    pub const FAILED_LOGIN_MAX_ATTEMPTS: u32 = 5;
 
     /// Grace period in seconds for refresh token reuse (to handle concurrent requests/retries).
     pub const REFRESH_TOKEN_GRACE_PERIOD_SECONDS: i64 = 10;
